clean-up comments
